add all tests
